create exceptions package
